+++ CHANGELOG.md
@@ -1,10 +1,11 @@
-# 1.0.4 (2014-04-23)
+# 1.0.4 (2014-05-11)
 
 - `string('')` now throws an error
 - `hex(upper)` now returns a cached function, for efficiency.
 - add `int32()`, which returns an integer from `-0x80000000` to `0x7fffffff`.
 - performance enhancements
 - add benchmark suite, covers engines, integer, bool, and real.
+- add `date(start, end)`, which returns a Date within a uniform distribution.
 
 # 1.0.3 (2014-04-23)